refactored component to use render method, fixed style for yt iframe
