+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/__tests__/quanticQuickviewContent.test.js
@@ -1,7 +1,7 @@
 import { createElement } from 'lwc';
-import QuanticQuickviewIframe from 'c/quanticQuickviewIframe';
+import QuanticQuickviewContent from 'c/quanticQuickviewContent';
 
-describe('c-quantic-quickview-iframe', () => {
+describe('c-quantic-quickview-content', () => {
     afterEach(() => {
         // The jsdom instance is shared across test cases in a single file so reset the DOM
         while (document.body.firstChild) {
@@ -11,8 +11,8 @@ describe('c-quantic-quickview-iframe', () => {
 
     it('TODO: test case generated by CLI command, please fill in test logic', () => {
         // Arrange
-        const element = createElement('c-quantic-quickview-iframe', {
-            is: QuanticQuickviewIframe
+        const element = createElement('c-quantic-quickview-content', {
+            is: QuanticQuickviewContent
         });
 
         // Act
