+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/quanticQuickviewContent.css
@@ -1,2 +1,3 @@
 @import 'c/quanticResultQuickview';
 @import 'c/quanticResultActionStyles';
+
