+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/quanticQuickviewContent.html
@@ -1,0 +1,2 @@
+<template>
+</template>
