+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/quanticQuickviewContent.js-meta.xml
