+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/quanticQuickviewDefault.html
@@ -1,0 +1,10 @@
+<template>
+  <!-- <iframe
+      id="quickview__content-container"
+      class="quickview__content-container slds-modal__content slds-p-around_large slds-wrap"
+      src={contentURL}
+      height="100%"
+      onload={onIframeLoaded}
+  ></iframe> -->
+  <div>hello</div>
+</template>
