+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/quanticQuickviewYoutube.css
@@ -1,0 +1,12 @@
+@import 'c/quanticResultQuickview';
+@import 'c/quanticResultActionStyles';
+
+.iframe-wrapper {
+  width: 100%;
+  height: 100%;
+  display: flex;
+}
+
+.iframe-wrapper iframe {
+  border: none;
+}
