--- packages/quantic/force-app/main/default/lwc/quanticQuickviewIframe/quanticQuickviewIframe.js
@@ -1,55 +1,0 @@
-import {
-  getHeadlessBundle,
-  getHeadlessEnginePromise,
-} from 'c/quanticHeadlessLoader';
-import {ResultUtils} from 'c/quanticUtils';
-import {LightningElement, api} from 'lwc';
-
-/** @typedef {import("coveo").Result} Result */
-/** @typedef {import("coveo").SearchEngine} SearchEngine */
-export default class QuanticQuickviewIframe extends LightningElement {
-  /**
-   * The ID of the engine instance the component registers to.
-   * @api
-   * @type {string}
-   */
-  @api engineId;
-  /**
-   * The [result item](https://docs.coveo.com/en/headless/latest/reference/search/controllers/result-list/#result).
-   * @api
-   * @type {Result}
-   */
-  @api result;
-
-  connectedCallback() {
-    getHeadlessEnginePromise(this.engineId)
-      .then((engine) => {
-        this.initialize(engine);
-      })
-      .catch((error) => {
-        console.error(error.message);
-      });
-  }
-
-  /**
-   * @param {SearchEngine} engine
-   */
-  initialize = (engine) => {
-    this.headless = getHeadlessBundle(this.engineId);
-    this.engine = engine;
-    ResultUtils.bindClickEventsOnResult(
-      this.engine,
-      this.result,
-      this.template,
-      this.headless.buildInteractiveResult
-    );
-  };
-
-  get youtubeIframeUrl() {
-    return 'https://www.youtube.com/embed/' + this.result?.raw?.ytvideoid;
-  }
-
-  get excerpts() {
-    return this.result?.excerpt;
-  }
-}
