+++ packages/quantic/force-app/main/default/lwc/quanticResultQuickview/quanticResultQuickview.js
@@ -274,24 +274,6 @@ export default class QuanticResultQuickview extends LightningElement {
     return !!this.previewButtonLabel;
   }
 
-  get contentURL() {
-    // console.log('result: ' + JSON.stringify(this.result));
-    // console.log('state: ' + JSON.stringify(this.state));
-    // console.log('raw: ' + JSON.stringify(this.result.raw.ytvideoid));
-    return this.state.contentURL?.includes(
-      encodeURIComponent(this.result.uniqueId)
-    )
-      ? this.state.contentURL
-      : undefined;
-  }
-
-  get isYoutube() {
-    if (this.result.uniqueId.includes('youtube')) {
-      return true;
-    }
-    return false;
-  }
-
   setFocusToHeader() {
     const focusTarget = this.template.querySelector('c-quantic-result-link');
